+++ inewave/nwlistop/modelos/hmont.py
@@ -1,18 +1,17 @@
 from inewave.config import MESES_DF
 
 from cfinterface.components.line import Line
-from cfinterface.components.literalfield import LiteralField
 from cfinterface.components.integerfield import IntegerField
 from cfinterface.components.floatfield import FloatField
 
-from inewave.nwlistop.modelos.blocos.valoresseriepatamar import (
-    ValoresSeriePatamar,
+from inewave.nwlistop.modelos.blocos.valoresserie import (
+    ValoresSerie,
 )
 
 
-class DtbmaxAnos(ValoresSeriePatamar):
+class HmontAnos(ValoresSerie):
     """
-    Bloco com as informações das tabelas de violação de turbinamento máximo
+    Bloco com as informações das tabelas de cota de montante
     da usina por mês/ano de estudo.
     """
 
@@ -20,7 +19,6 @@ class DtbmaxAnos(ValoresSeriePatamar):
     DATA_LINE = Line(
         [  # type: ignore
             IntegerField(4, 2),
-            LiteralField(5, 6),
         ]
-        + [FloatField(9, 14 + 9 * i, 2) for i in range(len(MESES_DF))]  # type: ignore
+        + [FloatField(15, 7 + 15 * i, 2) for i in range(len(MESES_DF))]  # type: ignore
     )